Lego WireLess Protocoll implementation proceeding
