+++ LegoBTLE/Device/SingleMotor.py
@@ -23,7 +23,7 @@
 # **************************************************************************************************
 from LegoBTLE.Device.ADevice import Device
 from LegoBTLE.Device.AMotor import AMotor
-from LegoBTLE.LegoWP.messages.downstream import DOWNSTREAM_MESSAGE_TYPE, DownStreamMessageBuilder
+from LegoBTLE.LegoWP.messages.downstream import DOWNSTREAM_MESSAGE_TYPE
 from LegoBTLE.LegoWP.messages.upstream import (DEV_GENERIC_ERROR, EXT_SERVER_MESSAGE, DEV_CMD_STATUS,
                                                DEV_PORT_NOTIFICATION,
                                                DEV_CURRENT_VALUE, HUB_ACTION, HUB_ATTACHED_IO)
@@ -34,7 +34,7 @@ class SingleMotor(AMotor, Device):
     
     def __init__(self,
                  name: str = 'SingleMotor',
-                 port: bytes=b'',
+                 port: bytes = b'',
                  gearRatio: float = 1.0,
                  debug: bool = False):
         
@@ -125,7 +125,6 @@ class SingleMotor(AMotor, Device):
     
     @property
     def current_cmd_snt(self) -> DOWNSTREAM_MESSAGE_TYPE:
-        
         return self._cmd_snt
 
     @property
